Merged PR 3154: Perform conf file write to link target instead of link file itself.
## Goals
* Ensure symlinks are preserved when used for the wpa_supplicant conf file.
* Ensure conf file changes are persisted across device updates.

## Technical Details
### Root Cause
wpa_supplicant updates its configuration file by writing out the entire configuration each time. This is done by first writing the config to a temporary file, which has the form <conf path>.tmp, then, renames this file to the true conf file path, <conf path> using the `rename` system call. The behavior of rename is such that if the destination file is a symlink, the symlink file itself will be replaced by the source file, _not_ the file the symlink target (necessarily so; otherwise you could not ever rename the link file itself).

### Fix
The fix is to ensure the symlink target is used as the path name. Since this path

## Reviewer Focus
* Security: the use of symlinks have historically been exploited by attackers to gain access to filesystem resources that were not otherwise authorized
* Think about how this might affect other config file access.

## Future Work
* We should consider if we want to ignore this problem and instead place the configuration file in a santa-cruz specific persisted location (/var). This would require separating the configuration out of the wpa_supplicant rpm package and placing it into either its own, or, as part of ztpd.
* Should also consider how the current santa cruz update process allows for persistence of config data normally stored on the rootfs, and whether or not some change would be helpful to avoid the use of symlinks.