Merged PR 3197: Fix wpas getting stuck in 'authenticated' state after removing networks.
## Goals
* Fix the case where an enrollee gets provisioned, then its network is manually removed and doesn't resume chirping.
* Allow an enrollee to be re-provisioned continuously when networks are removed in any way.

## Technical Details
ztpd listens for dpp state changes from wpa_supplicant to determine its next action. When authentication completes, the dpp state transitions to 'authenticated'. However, when networks are removed, the state isn't also updated.

The dpp state should not encode transient states that don't have actions related to them. Consequently, the state itself should be re-encoded with fewer states that only describe actions: inactive, chirping. This removal will be done in a future change when upstream changes are merged.

For this PR, the dpp state will transition to 'inactive' which describes that chirping is not occurring, triggering ztpd to re-check if there are configured networks, and kick-off chirping again.

## Reviewer Focus
* Consider how this might affect the ztpd state machine
* Consider how this might affect wpa_supplicant DPP logic

## Future Work
* The `enum dpp_state` will either be pruned of ineffectual entries, or removed all-together and replaced with a 'chirping' boolean signal.