Use wpa conf file link target for writing.
When the wpa conf file is a symbolic link, the link is clobbered when
written. This occurs because the rename() of the (regular) tmp file
used to write changes replaces the original file, thus discarding the
link.

Fix this by first resolving the conf file link target (if any), and
using the target as the defacto file path instead of the link file
itself. This ensures the tmp file is created on the same filesystem
as the link target, which is required for rename().

Signed-off-by: Andrew Beltrano <anbeltra@microsoft.com>