Generate an event when a network is added or removed.
Generate an event on the control socket interface when a network is added or
removed. The event name CTRL-EVENT-NETWORK-<ADDED|REMOVED> is followed by the
network entry identifier and a unary string argument 'last' if the removed
network was the last entry remaining. The event matches the corresponding
Network<Added|Removed> signal on the d-bus interface.

Signed-off-by: Andrew Beltrano <anbeltra@microsoft.com>